Add utility functions for handling list of choices
